feat: navbar, category ui변경
