+++ frontend/src/pages/LandingPage/Sections/PriceCheckBox.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const PriceCheckBox = ({ price, checkedPrice, onFilters }) => {
   return (
-    <div className='p-2 mb-3 bg-blue-300 rounded-md'>
+    <div className='p-3'>
       {price?.map(item => (
         <div key={item._id}>
           <input
