+++ frontend/src/pages/LandingPage/index.jsx
@@ -5,6 +5,12 @@ import CardItem from './Sections/CardItem';
 import axiosInstance from '../../utils/axios';
 import DepartmentCheckBox from './Sections/DepartmentCheckBox';
 import { department, price } from '../../utils/filterData';
+import { Fragment } from 'react';
+import { Menu, Transition } from '@headlessui/react';
+import { ChevronDownIcon } from '@heroicons/react/20/solid';
+// function classNames(...classes) {
+//   return classes.filter(Boolean).join(' ');
+// }
 const LandingPage = () => {
   const limit = 4;
   const [searchTerm, setSearchTerm] = useState('');
@@ -102,24 +108,65 @@ const LandingPage = () => {
   };
   return (
     <section>
-      <div className='text-center m-7'>
-        <h2 className='text-2xl'>중고 상품 페이지</h2>
-      </div>
       {/* filter */}
-      <div className='flex gap-3'>
-        <div className='w-1/2'>
-          <DepartmentCheckBox
-            department={department}
-            checkedDepartment={filters.department}
-            onFilters={filters => handleFilters(filters, 'department')}
-          />
+      <div className='flex gap-[300px] bg-gray-100 justify-center p-7 mb-10'>
+        <div>
+          <Menu as='div' className='relative inline-block text-left'>
+            <div>
+              <Menu.Button className='inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50'>
+                학과 카테고리
+                <ChevronDownIcon className='-mr-1 h-5 w-5 text-gray-400' aria-hidden='true' />
+              </Menu.Button>
+            </div>
+
+            <Transition
+              as={Fragment}
+              enter='transition ease-out duration-100'
+              enterFrom='transform opacity-0 scale-95'
+              enterTo='transform opacity-100 scale-100'
+              leave='transition ease-in duration-75'
+              leaveFrom='transform opacity-100 scale-100'
+              leaveTo='transform opacity-0 scale-95'>
+              <Menu.Items className='absolute right-0 z-10 mt-2 w-56 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none'>
+                <div className='py-1'>
+                  <DepartmentCheckBox
+                    department={department}
+                    checkedDepartment={filters.department}
+                    onFilters={filters => handleFilters(filters, 'department')}
+                  />
+                </div>
+              </Menu.Items>
+            </Transition>
+          </Menu>
         </div>
-        <div className='w-1/2'>
-          <PriceCheckBox
-            price={price}
-            checkedPrice={filters.price}
-            onFilters={filters => handleFilters(filters, 'price')}
-          />
+        <div>
+          <Menu as='div' className='relative inline-block text-left'>
+            <div>
+              <Menu.Button className='inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50'>
+                가격 카테고리
+                <ChevronDownIcon className='-mr-1 h-5 w-5 text-gray-400' aria-hidden='true' />
+              </Menu.Button>
+            </div>
+
+            <Transition
+              as={Fragment}
+              enter='transition ease-out duration-100'
+              enterFrom='transform opacity-0 scale-95'
+              enterTo='transform opacity-100 scale-100'
+              leave='transition ease-in duration-75'
+              leaveFrom='transform opacity-100 scale-100'
+              leaveTo='transform opacity-0 scale-95'>
+              <Menu.Items className='absolute right-0 z-10 mt-2 w-56 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none'>
+                <div className='py-1'>
+                  <PriceCheckBox
+                    price={price}
+                    checkedPrice={filters.price}
+                    onFilters={filters => handleFilters(filters, 'price')}
+                  />
+                </div>
+              </Menu.Items>
+            </Transition>
+          </Menu>
         </div>
       </div>
       {/* search */}
