+++ frontend/src/pages/NavTestPage/index.jsx
@@ -1,168 +1,51 @@
-import { Fragment, useState } from 'react';
-import { Disclosure, Menu, Transition } from '@headlessui/react';
-import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
-import React from 'react';
-const navigation = [
-  { name: '홈', href: '/' },
-  { name: '업로드', href: 'product/upload' },
-  { name: '주문목록', href: '#' },
-  { name: '문의하기', href: '#' },
-];
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
+// import React, { useState } from 'react';
+// import { Link } from 'react-router-dom';
+// import NavItems from './Sections/NavItems';
+// // import { Fragment } from 'react';
+// // import { Disclosure, Menu, Transition } from '@headlessui/react';
+// // import { Bars3Icon, BellIcon, XMarkIcon } from '@heroicons/react/24/outline';
+// // const navigation = [
+// //   { name: '홈으로', href: '/', current: true },
+// //   { name: '업로드', href: 'product/upload', current: false },
+// //   { name: '주문목록', href: '#', current: false },
+// //   // { name: 'Calendar', href: '#', current: false },
+// // ];
+// // function classNames(...classes) {
+// //   return classes.filter(Boolean).join(' ');
+// // }
+// const Navbar = () => {
+//   const [menu, setMenu] = useState(false);
+//   const handleMenu = () => {
+//     setMenu(!menu);
+//   };
+//   return (
+//     <nav className='relative z-10 bg-white-300'>
+//       <div className='w-full h-30'>
+//         <div className='flex items-center justify-between mx-5 sm:mx-10 lg:mx-20'>
+//           {/* logo */}
+//           <div className='flex items-center text-2xl h-30'>
+//             <Link to='/'>
+//               <img src={'/image/mainicon.png'} alt={'아이콘'} />
+//             </Link>
+//           </div>
 
-const NavTestPage = () => {
-  // const changeCurrentStatus = e => {};
-  const [select, setSelect] = useState('홈');
-  const handleClick = name => {
-    setSelect(name);
-  };
-  return (
-    <Disclosure as='nav' className='bg-gray-800'>
-      {({ open }) => (
-        <>
-          <div className='mx-auto max-w-7xl px-2 sm:px-6 lg:px-8'>
-            <div className='relative flex h-16 items-center justify-between'>
-              <div className='absolute inset-y-0 left-0 flex items-center sm:hidden'>
-                {/* Mobile menu button*/}
-                <Disclosure.Button className='inline-flex items-center justify-center rounded-md p-2 text-gray-400 hover:bg-gray-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white'>
-                  <span className='sr-only'>Open main menu</span>
-                  {open ? (
-                    <XMarkIcon className='block h-6 w-6' aria-hidden='true' />
-                  ) : (
-                    <Bars3Icon className='block h-6 w-6' aria-hidden='true' />
-                  )}
-                </Disclosure.Button>
-              </div>
-              <div className='flex flex-1 items-center justify-center sm:items-stretch sm:justify-start'>
-                <div className='flex flex-shrink-0 items-center'>
-                  <img
-                    className='block h-10 w-auto lg:hidden'
-                    src={'/image/mainicon.png'}
-                    alt='Your Company'
-                  />
-                  <img
-                    className='hidden h-10 w-auto lg:block'
-                    src={'/image/mainicon.png'}
-                    alt='Your Company'
-                  />
-                </div>
-                <div className='hidden sm:ml-6 sm:block'>
-                  <div className='flex space-x-4'>
-                    {navigation.map(item => (
-                      <a
-                        key={item.name}
-                        href={item.href}
-                        className={classNames(
-                          item.name === select
-                            ? 'bg-gray-900 text-white'
-                            : 'text-gray-300 hover:bg-gray-700 hover:text-white',
-                          'rounded-md px-3 py-2 text-sm font-medium'
-                        )}
-                        aria-current={item.name === select ? 'page' : undefined}
-                        onClick={() => handleClick(item.name)}>
-                        {item.name}
-                      </a>
-                    ))}
-                  </div>
-                </div>
-              </div>
-              <div className='absolute inset-y-0 right-0 flex items-center pr-2 sm:static sm:inset-auto sm:ml-6 sm:pr-0'>
-                {/* <button
-                  type='button'
-                  className='rounded-full bg-gray-800 p-1 text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800'>
-                  <span className='sr-only'>View notifications</span>
-                  <BellIcon className='h-6 w-6' aria-hidden='true' />
-                </button> */}
+//           {/* menu button */}
+//           <div className='text-2xl sm:hidden'>
+//             <button onClick={handleMenu}>{menu ? '-' : '+'}</button>
+//           </div>
 
-                {/* Profile dropdown */}
-                <Menu as='div' className='relative ml-3'>
-                  <div>
-                    <Menu.Button className='flex rounded-full bg-gray-800 text-sm focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-gray-800'>
-                      <span className='sr-only'>Open user menu</span>
-                      <img
-                        className='h-8 w-8 rounded-full'
-                        src='https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
-                        alt=''
-                      />
-                    </Menu.Button>
-                  </div>
-                  <Transition
-                    as={Fragment}
-                    enter='transition ease-out duration-100'
-                    enterFrom='transform opacity-0 scale-95'
-                    enterTo='transform opacity-100 scale-100'
-                    leave='transition ease-in duration-75'
-                    leaveFrom='transform opacity-100 scale-100'
-                    leaveTo='transform opacity-0 scale-95'>
-                    <Menu.Items className='absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none'>
-                      <Menu.Item>
-                        {({ active }) => (
-                          <a
-                            href='#'
-                            className={classNames(
-                              active ? 'bg-gray-100' : '',
-                              'block px-4 py-2 text-sm text-gray-700'
-                            )}>
-                            Your Profile
-                          </a>
-                        )}
-                      </Menu.Item>
-                      <Menu.Item>
-                        {({ active }) => (
-                          <a
-                            href='#'
-                            className={classNames(
-                              active ? 'bg-gray-100' : '',
-                              'block px-4 py-2 text-sm text-gray-700'
-                            )}>
-                            Settings
-                          </a>
-                        )}
-                      </Menu.Item>
-                      <Menu.Item>
-                        {({ active }) => (
-                          <a
-                            href='#'
-                            className={classNames(
-                              active ? 'bg-gray-100' : '',
-                              'block px-4 py-2 text-sm text-gray-700'
-                            )}>
-                            Sign out
-                          </a>
-                        )}
-                      </Menu.Item>
-                    </Menu.Items>
-                  </Transition>
-                </Menu>
-              </div>
-            </div>
-          </div>
+//           {/* big screen */}
+//           <div className='hidden sm:block'>
+//             {' '}
+//             <NavItems />
+//           </div>
+//         </div>
 
-          <Disclosure.Panel className='sm:hidden'>
-            <div className='space-y-1 px-2 pb-3 pt-2'>
-              {navigation.map(item => (
-                <Disclosure.Button
-                  key={item.name}
-                  as='a'
-                  href={item.href}
-                  className={classNames(
-                    item.current
-                      ? 'bg-gray-900 text-white'
-                      : 'text-gray-300 hover:bg-gray-700 hover:text-white',
-                    'block rounded-md px-3 py-2 text-base font-medium'
-                  )}
-                  aria-current={item.current ? 'page' : undefined}>
-                  {item.name}
-                </Disclosure.Button>
-              ))}
-            </div>
-          </Disclosure.Panel>
-        </>
-      )}
-    </Disclosure>
-  );
-};
+//         {/* mobile */}
+//         <div className='block sm:hidden'>{menu && <NavItems mobile />}</div>
+//       </div>
+//     </nav>
+//   );
+// };
 
-export default NavTestPage;
+// export default Navbar;
